ONU scalar values can now be floats; The buffer between the sleep time and the actual time slept is now a fraction of the wakeup time instead of the sleep time; coo.txt contains the count of the error instances

+++ code/eponsim.h
@@ -210,7 +210,7 @@ typedef struct
   double    ACTUAL_MIN_PROP_DELAY;
   double    ONU_PROP[MAX_ONU];    /* ONU to OLT propagation delay values */
 
-  int       ONUTrafficScalar[MAX_ONU];
+  double    ONUTrafficScalar[MAX_ONU];
   double    ONUTrafficScalar_link_speed[MAX_ONU];
   int       ONUTrafficScalar_length;
   double    AVG_PKT_INTER_ARVL_TIME_ONU[MAX_ONU];
@@ -435,9 +435,6 @@ extern sONU_LIST *scheduleList;
 /* Scheduling Pool structure array */
 extern      sSCHED_POOL schedPool[MAX_ONU];
 extern int  schedPoolCount;
-
-/* Array of weights for changing the traffic loads between ONUs */
-//extern int  ONUTrafficScalar[MAX_ONU];
 
 /*
  * Function declarations
